transport: conn, add QueryAsync
BenchmarkConnQueryIntegration-8             2233           2654839 ns/op            1251 B/op         34 allocs/op
BenchmarkConnAsyncQueryIntegration-8       10000           2088146 ns/op            1289 B/op         11 allocs/op

- QueryAsync runs Query in a separated goroutine
- Benchmarks moved to separate file
- connTestHelper added applyFixture
- requestChanSize increased to maxStreamID/2, with 1k goroutines block on putting to the channel
- newQueryForStatement changed to makeQueryForStatement returning Query this removes the gc pressure
